% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MesFonctions-package.r
\docType{package}
\name{MesFonctions}
\alias{MesFonctions}
\alias{MesFonctions-package}
\title{MesFonctions: A collection of functions for linear models}
\description{
MesFonctions: A collection of functions for linear models
}
