% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRESS.R
\name{PRESS}
\alias{PRESS}
\title{PRESS Function}
\usage{
PRESS(model, ...)
}
\arguments{
\item{model}{An lm object. (Requires input)}
}
\description{
This function computes Allen's PRESS for an lm object
}
\examples{
x<-runif(1:20); y<-2+3*x+rnorm(20,0,5); fit<-lm(y~x)
PRESS(fit)
x<-runif(1:20); y<-2+3*x+rnorm(20,0,5); fit2<-lm(y~x)
PRESS(fit,fit2)
}
\keyword{PRESS}
