% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSS.R
\name{RSS}
\alias{RSS}
\title{RSS Function}
\usage{
RSS(model)
}
\arguments{
\item{model}{An lm object. (Requires input)}
}
\description{
This function computes the residual sum of squares for an lm object
}
\examples{
x<-runif(1:20); y<-2+3*x+rnorm(20,0,5); fit<-lm(y~x)
RSS(fit)
}
\keyword{RSS}
